# services/registry-api/openapi.yaml
openapi: 3.1.0
info:
  title: AKL Identity & Document Registry API
  description: Registry service for controlled documents, versions, access policies,
    authorization decisions, and audit events. It does not implement ingestion, RAG,
    or LLM logic.
  version: 0.1.0
paths:
...
  /api/v1/documents:
    post:
      summary: Create Document
      operationId: create_document_api_v1_documents_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      summary: List Documents
      operationId: list_documents_api_v1_documents_get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/DocumentStatus'
          - type: 'null'
          title: Status
      - name: classification
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Classification'
          - type: 'null'
          title: Classification
      - name: document_type
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/DocumentType'
          - type: 'null'
          title: Document Type
      - name: owner_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner Id
      - name: tag
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Tag
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 100
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/documents/{document_id}:
    get:
      summary: Get Document
      operationId: get_document_api_v1_documents__document_id__get
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          title: Document Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      summary: Patch Document
      operationId: patch_document_api_v1_documents__document_id__patch
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          title: Document Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentPatch'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      summary: Delete Document
      operationId: delete_document_api_v1_documents__document_id__delete
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          title: Document Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/documents/{document_id}/versions:
    post:
      summary: Create Document Version
      operationId: create_document_version_api_v1_documents__document_id__versions_post
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          title: Document Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentVersionCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentVersionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      summary: List Document Versions
      operationId: list_document_versions_api_v1_documents__document_id__versions_get
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          title: Document Id
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/DocumentStatus'
          - type: 'null'
          title: Status
      - name: valid_on
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Valid On
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 100
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentVersionListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/documents/{document_id}/versions/{version_id}:
    get:
      summary: Get Document Version
      operationId: get_document_version_api_v1_documents__document_id__versions__version_id__get
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          title: Document Id
      - name: version_id
        in: path
        required: true
        schema:
          type: string
          title: Version Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentVersionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/documents/{document_id}/versions/{version_id}/publish:
    post:
      summary: Publish Document Version
      operationId: publish_document_version_api_v1_documents__document_id__versions__version_id__publish_post
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          title: Document Id
      - name: version_id
        in: path
        required: true
        schema:
          type: string
          title: Version Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentVersionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/documents/{document_id}/versions/{version_id}/archive:
    post:
      summary: Archive Document Version
      operationId: archive_document_version_api_v1_documents__document_id__versions__version_id__archive_post
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          title: Document Id
      - name: version_id
        in: path
        required: true
        schema:
          type: string
          title: Version Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentVersionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/authz/check:
    post:
      summary: Check Authorization
      operationId: check_authorization_api_v1_authz_check_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthzCheckRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthzCheckResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/authz/filter-documents:
    post:
      summary: Filter Authorized Documents
      operationId: filter_authorized_documents_api_v1_authz_filter_documents_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthzFilterDocumentsRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthzFilterDocumentsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/audit/events/{event_id}:
    get:
      summary: Get Audit Event
      operationId: get_audit_event_api_v1_audit_events__event_id__get
      parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: string
          title: Event Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEventResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    DocumentCreate:
      properties:
        title:
          type: string
          maxLength: 300
          minLength: 1
          title: Title
        document_type:
          $ref: '#/components/schemas/DocumentType'
        owner_id:
          type: string
          maxLength: 128
          minLength: 1
          title: Owner Id
        gestor_unit:
          anyOf:
          - type: string
            maxLength: 128
          - type: 'null'
          title: Gestor Unit
        classification:
          $ref: '#/components/schemas/Classification'
          default: internal
        tags:
          items:
            type: string
          type: array
          title: Tags
        metadata:
          additionalProperties: true
          type: object
          title: Metadata
        access_policies:
          anyOf:
          - items:
              $ref: '#/components/schemas/AccessPolicyCreate'
            type: array
          - type: 'null'
          title: Access Policies
      type: object
      required:
      - title
      - document_type
      - owner_id
      title: DocumentCreate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...